Text Summarization/requirements.txt
langchain
python-dotenv
langchain_community
wikipedia
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
ipykernel
validators==0.28.1
youtube_transcript_api
pytube
unstructured